refactor(backend): move export filename builder into the domain core
buildExportFilename is a pure function, not a use case, and it was the only
non-Injectable file under any use-cases directory in the backend. Move it to
experiments/core/ alongside its spec, matching visibility/core/visibility-
transition.ts and sharing/core/grantee-selectability.ts, which are the
established home for pure domain rules consumed by use cases.

Also reject impossible calendar dates in parseDatabricksTimestamp. Date rolls
"2026-02-30" forward to 2026-03-02, which would have put a plausible but wrong
timestamp in the filename instead of falling back to the export id.
