+++ apps/backend/src/common/utils/datetime.spec.ts
@@ -28,6 +28,18 @@ describe("parseDatabricksTimestamp", () => {
     );
   });
 
+  it("rejects impossible calendar dates instead of rolling them forward", () => {
+    expect(parseDatabricksTimestamp("2026-02-30 03:04:05")).toBeNull();
+    expect(parseDatabricksTimestamp("2026-13-01 03:04:05")).toBeNull();
+    expect(parseDatabricksTimestamp("2026-01-32 03:04:05")).toBeNull();
+  });
+
+  it("still accepts real leap days", () => {
+    expect(parseDatabricksTimestamp("2024-02-29 03:04:05")?.toISOString()).toBe(
+      "2024-02-29T03:04:05.000Z",
+    );
+  });
+
   it("returns null for empty or unparsable values", () => {
     expect(parseDatabricksTimestamp(null)).toBeNull();
     expect(parseDatabricksTimestamp(undefined)).toBeNull();
