+++ apps/backend/src/common/utils/datetime.ts
@@ -9,9 +9,12 @@ export function compactTimestamp(date: Date = new Date()): string {
   return `${datePart}_${timePart}`;
 }
 
+const OFFSETLESS = /^(\d{4}-\d{2}-\d{2})[ T](\d{2}:\d{2}:\d{2})(?:\.\d+)?$/;
+
 /**
  * Databricks SQL timestamps carry no offset, so treat those as UTC instead of
- * letting the parse drift with the server's timezone.
+ * letting the parse drift with the server's timezone. Rejects impossible
+ * calendar dates, which `Date` would otherwise roll forward into a wrong instant.
  */
 export function parseDatabricksTimestamp(value: string | null | undefined): Date | null {
   const normalized = value?.trim();
@@ -19,10 +22,16 @@ export function parseDatabricksTimestamp(value: string | null | undefined): Date
     return null;
   }
 
-  const withOffset = /^\d{4}-\d{2}-\d{2}[ T]\d{2}:\d{2}:\d{2}(?:\.\d+)?$/.test(normalized)
-    ? `${normalized.replace(" ", "T")}Z`
-    : normalized;
+  const offsetless = OFFSETLESS.exec(normalized);
+  const date = new Date(offsetless ? `${normalized.replace(" ", "T")}Z` : normalized);
+  if (Number.isNaN(date.getTime())) {
+    return null;
+  }
 
-  const date = new Date(withOffset);
-  return Number.isNaN(date.getTime()) ? null : date;
+  // "2026-02-30" parses as 2026-03-02; only a round trip catches the rollover.
+  if (offsetless && !date.toISOString().startsWith(`${offsetless[1]}T${offsetless[2]}`)) {
+    return null;
+  }
+
+  return date;
 }
