+++ apps/backend/src/experiments/application/use-cases/experiment-data-exports/download-export.ts
@@ -3,9 +3,9 @@ import { Readable } from "stream";
 
 import { ErrorCodes } from "../../../../common/utils/error-codes";
 import { Result, AppError, failure, success } from "../../../../common/utils/fp-utils";
+import { buildExportFilename } from "../../../core/export-filename";
 import { ExperimentDataExportsRepository } from "../../../core/repositories/experiment-data-exports.repository";
 import { ExperimentRepository } from "../../../core/repositories/experiment.repository";
-import { buildExportFilename } from "./export-filename";
 
 /**
  * Use case for downloading a completed export by export ID
